requirements.txt
asgiref==3.3.1 
django==3.1.6 
pytz==2021.1 
sqlparse==0.4.1
djangorestframework==3.12.2
cffi==1.14.5 
cryptography==3.4.4 
django-rest-knox==4.1.0 
pycparser==2.20
npm==0.1.1 
optional-django==0.1.0
ippanel==1.1.0
psycopg2==2.8.6
Pillow==8.2.0
#add schedule